+++ lib/prepare.dart
@@ -1,0 +1,11 @@
+library;
+
+export 'command/prepare_build_command.dart';
+export 'command/prepare_watch_command.dart';
+
+export 'components/prepare_builder.dart';
+export 'components/prepare_queue.dart';
+export 'components/prepare_debounce.dart';
+export 'components/source_file.dart';
+export 'components/source_file_edit.dart';
+export 'components/source_file_manager.dart';